backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
sqlalchemy>=2.0.35
python-dotenv>=1.0.1
httpx>=0.27.0
websockets>=13.0
